feat(loom): workspace-wide loom compile + runtime support
Make `just features=loom test` build and run end-to-end across the
workspace. Earlier commits had loom working only on the `concurrency`
crate; everything else failed to compile or crashed at runtime with
stack overflow, Arc leak, or DashMap destructor panics. Bundled into
one commit because each fix was discovered by running the previous
one.

## concurrency

  * `fn sleep(_: Duration)` shim under loom (loom 0.7 doesn't model
    time; yields to the scheduler so the call still acts as a
    schedule point).
  * Enumerate `loom::thread` re-exports and shadow `spawn` with a
    4 MiB default stack. Loom 0.7's default coroutine stack is 4 KiB,
    which overflows trivially under atomic-heavy `concurrency::sync`
    types.
  * `stress` under loom wraps the body in a 4 MiB
    `Builder::spawn` so the main 4 KiB coroutine just spawns and
    joins. Costs one of loom's five thread slots.
  * `loom_scope::Scope::spawn` routes through `super::spawn` for the
    same default.
  * `Slot::load()` / `SlotOption::load()` helpers used by
    `common::cliprovider` to drop a redundant `load_full()` clone.

## nat

  * `cfg_attr`-gate `#![feature(arbitrary_self_types)]`: loom wraps
    `Arc<T>` in a facade newtype that isn't a blessed self-receiver,
    so `self: Arc<Self>` methods on `AllocatedIp` /
    `AllocatedPortBlock` need the unstable feature there.
  * `#[concurrency_mode(loom)]` no-op `shuffle_slice` (loom needs
    determinism for replay; shuffle is allocation-order heuristic,
    not correctness).
  * Gate `concurrency_tests` off loom: the facade's `Weak` shim holds
    a strong clone of the `Arc`, so the allocator's
    `Weak::upgrade().is_none()` liveness signal never fires and
    loom's `Arc leaked` assertion catches it.

## flow-entry

  * Gate `concurrency_tests` to shuttle only: `FlowTable`'s internal
    `DashMap` panics in loom's end-of-execution cleanup (sharded
    `RwLock`s don't fit loom's strict lifecycle accounting).

## routing

  * Gate `fib::test::concurrency_tests` off loom: the `left_right`
    epoch state space is too large for exhaustive search to terminate
    in reasonable time.

## dataplane

  * The binary builds an `Arc<dyn Fn(...) ...>` trait-object closure
    in `packet_processor::setup_internal`, which needs `CoerceUnsized`
    on the concrete `Arc`. Loom 0.7's `Arc` doesn't carry that trait,
    and the facade newtype can't add it. Gate the bin out of loom
    builds: extract the body to `dataplane/src/runtime.rs` and leave
    `main.rs` with a stub `main` under loom that panics if invoked.
    Library crates still get loom coverage through feature
    propagation.
  * `dataplane/src/drivers/dpdk.rs` switches `use crate::CmdArgs` to
    `use args::CmdArgs;` to follow the new module layout.

CI's loom step in `.github/workflows/dev.yml` still scopes to
`--package=dataplane-concurrency` because the loom-incompatible tests
across the workspace are now cfg-gated rather than package-filtered;
the package scope is no longer load-bearing for the test invocation,
only for cargo's feature unification.

Default and `--features shuttle` builds unaffected.
